# Generate addition chains for exponentiation by fixed value in `Sqrt`