--- a/.gitignore
+++ b/.gitignore
@@ -4,6 +4,7 @@
 *.dll
 *.so
 *.dylib
+*.txt
 
 # Test binary, build with `go test -c`
 *.test