@gbotrel 2021-11-11
Update: worth it to keep it for `Legendre` (similar perf improvement). 

Keeping it for sake of code generation simplicity for bls12-377, but this #98 will change our numbers anyway.